+++ src/Opaleye/Internal/Column.hs
@@ -7,10 +7,7 @@ import qualified Opaleye.Internal.HaskellDB.PrimQuery as HPQ
 -- | A column of a @Query@, of type @pgType@.  For example 'Column'
 -- @PGInt4@ is an @int4@ column and a 'Column' @PGText@ is a @text@
 -- column.
---
--- Do not use the 'Show' instance of 'Column'.  It is considered
--- deprecated and will be removed in version 0.7.
-newtype Column pgType = Column HPQ.PrimExpr deriving Show
+newtype Column pgType = Column HPQ.PrimExpr
 
 -- | Only used within a 'Column', to indicate that it can be @NULL@.
 -- For example, a 'Column' ('Nullable' @PGText@) can be @NULL@ but a