+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.7.0.0
+
+* Removed deprecated functionality
+
+  * `runInsert`, `runInsertReturning`, `showPGType`, `literalColumn`,
+    `unsafePgFormatTime`, `showSqlForPostgres`,
+    `showSqlForPostgresUnopt`, `showSqlForPostgresExplicit`,
+    `showSqlForPostgresUnoptExplicit`, `formatAndShowSQL`,
+    `prepareQuery`, `Show` instance of `Column`, `View`, `Writer`,
+    `Table` (in favour of `table`), `TableWithSchema` (in favour of
+    `tableWithSchema`)
+
+* Deprecated
+
+  * `tableProperties`, `exists`, `notExists`
+
 ## 0.6.0.0
 
 * Added `runUpdateEasy`